update value of var Quantity in Product

+++ Infrastructure/ViewModel/Request/ProductRequest.cs
@@ -67,7 +67,7 @@ namespace Application.ViewModel.Request
         {
             [Required(ErrorMessage = "Stock Quantity is required.")]
             [Range(0, int.MaxValue, ErrorMessage = "Stock Quantity cannot be negative.")]
-            public int? StockQuantity { get; set; }
+            public double? StockQuantity { get; set; }
         }
     }
 }
